+++ src/main/java/com/labsky/timotej/model/products/ProductFactory.java
@@ -4,6 +4,8 @@ import com.labsky.timotej.model.constants.ProductTypes;
 import com.labsky.timotej.model.constants.ProductsFileFields;
 import com.labsky.timotej.model.products.promotions.InsuranceDiscount;
 
+import java.math.BigDecimal;
+
 import static java.lang.Double.parseDouble;
 
 /**
@@ -17,23 +19,27 @@ public class ProductFactory {
         return switch (productFields[ProductsFileFields.TYPE]) {
             case ProductTypes.GENERIC_PRODUCT -> GenericProduct.builder()
                     .name(productFields[ProductsFileFields.NAME])
-                    .price(parseDouble(productFields[ProductsFileFields.PRICE]))
+                    .price(BigDecimal.valueOf(
+                            parseDouble(productFields[ProductsFileFields.PRICE])))
                     .currency(productFields[ProductsFileFields.CURRENCY])
                     .build();
             case ProductTypes.SIM -> SimCard.builder()
                     .name(productFields[ProductsFileFields.NAME])
-                    .price(parseDouble(productFields[ProductsFileFields.PRICE]))
+                    .price(BigDecimal.valueOf(
+                            parseDouble(productFields[ProductsFileFields.PRICE])))
                     .currency(productFields[ProductsFileFields.CURRENCY])
                     .build();
             case ProductTypes.INSURANCE -> Insurance.builder()
                     .name(productFields[ProductsFileFields.NAME])
-                    .price(parseDouble(productFields[ProductsFileFields.PRICE]))
+                    .price(BigDecimal.valueOf(
+                            parseDouble(productFields[ProductsFileFields.PRICE])))
                     .currency(productFields[ProductsFileFields.CURRENCY])
                     .salePromotions(new InsuranceDiscount())
                     .build();
             case ProductTypes.EARPHONES -> Earphones.builder()
                     .name(productFields[ProductsFileFields.NAME])
-                    .price(parseDouble(productFields[ProductsFileFields.PRICE]))
+                    .price(BigDecimal.valueOf(
+                            parseDouble(productFields[ProductsFileFields.PRICE])))
                     .currency(productFields[ProductsFileFields.CURRENCY])
                     .build();
             default -> null;
